pyproject.toml
[project]
name = "crop_doctor_env"
version = "1.0.0"
description = "AI Agricultural Disease Diagnosis — an OpenEnv RL environment"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.3",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
]

[project.scripts]
server = "server.app:main"
